+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Reset the validator's schema at every call in the ODS configuration example.
 - Update the ifm3d playground to use the `Port` function to get the PCIC port.
 - Update the Python `update_settings_to_new_fw_schema.py` example to create a `logs` directory if it does not already exist.
+- Fix `ssh_key_gen.py` overwrites authorized keys rather than append to them.
 
 ## 1.2.0
 - Update the examples for O3D3xx and O3X1xx for ifm3d 1.5.3.